chore: organize some functions
