+++ src/env.h
@@ -1,0 +1,6 @@
+#ifndef ENV_H_ /* Include guard */
+#define ENV_H_
+
+int getVacuumCount();
+
+#endif // ENV_H_
