+++ src/extent.h
@@ -1,0 +1,12 @@
+#ifndef EXTENT_H_ /* Include guard */
+#define EXTENT_H_
+
+typedef struct Extent
+{
+    double minX;
+    double minY;
+    double maxX;
+    double maxY;
+} Extent;
+
+#endif // EXTENT_H_
