+++ src/statement.h
@@ -5,6 +5,7 @@
 #include <stdlib.h>
 #include <string.h>
 #include <sqlite3.h>
+#include "extent.h"
 
 sqlite3_stmt *prepareStatement(sqlite3 *db, char *query, int flags);
 
@@ -32,6 +33,12 @@ void finalizeStatement(sqlite3_stmt *stmt);
 void vacuum(sqlite3 *db);
 
 int getBlobSize(sqlite3 *db, sqlite3_stmt *stmt, char *tileCache, int z, int x, int y);
+
+char *getAddIndexQuery(char *tileCache);
+
+char *getTileCountQuery(char *tileCache);
+
+sqlite3_stmt *getExtentInsertStmt(sqlite3 *db, Extent *extent);
 
 sqlite3_stmt *getBatchSelectStmt(sqlite3 *db, char *tileCache);
